initial commit of UF-Algebra and last commit before major overhaul
I've decided we should revise our implementation of subuniverse/subalgebra in
UF-Subuniverse.agda and base it on MHE's development of subgroups (see the
current version of UF-Structures.agda).

So, we will generalize what is now in UF-Structures.agda so that subalgebras are defined for
general algebras (not just for groups).