some major refactoring involving homs
The UF-Hom.agda file now distinguishes between intensional/extensional versions
of hom conditions.  I thought this would enable us to avoid using functional
extensionality in some of our proofs... and it does... but it may be more
trouble than it's worth.

Please use hom, instead of Hom, to refer to an (intensional) homomorphism (which
is our old, standard notion of a hom).

We now use the following notation:

  * HOM = functions that satisfy the fully intensional hom condition
  * Hom = functions that satisfy the partially intensional hom condition
  * hom = functions that satisfy the extensional hom condition

See UF-Hom.agda for details.