make indexing type slightly more general
We now let the indexing type live at the same level as the things being indexed.

If we try to make this more general by using, say, I : Type ι, then the
intersection will live at a higher universe than the conjucts (the ci's),
so we would have to resort to using Lift.