initial commit of substructures module
The structures.substructures.basic module defines types representing subuniverses
of structures and facts about them. Next we'll add a structures.substructures.substructures
module for types representing structures that are substructures of other
structures.

For subuniverses, not much has changed from the analogous development for
algebras.  The relations play no real role since subuniverses are still just
closed under the operations.  However, once we move on to substructures,
relations will play a role since the relations of substructures will be the
relations of the parent structure restricted to the universe of the
substructure.