new approach to proving swelldef in MLTT
We now have a proof of swelldef in the special case of binary operations
(of type A -> A -> A) and slightly more generally for binary functions
(of type A -> A -> B).

Next, we hope to generalize this to (Fin 2 -> A) -> B, and then to
(Fin n -> A) -> B for arbitrary n in Nat.

If that works, then we should determine whether the approach works in complete
generality (i.e., for functions of type (I -> A) -> B).